Add files via upload
This file, named "inpFle_BCCP.csv", is a comma-separated values (CSV) file containing key financial metrics for credit card customers. It serves as input data for the Supervised Learning Models. The data includes the following attributes:

credit_score: The credit score of the customer.
isCtry1: A binary variable indicating the customer's country (1 for country 1, 0 otherwise).
isCtry2: A binary variable indicating the customer's country (1 for country 2, 0 otherwise).
isFemale: A binary variable indicating the customer's gender (1 for female, 0 for male).
age: The customer's age in years.
tenure: The customer's tenure with the credit card company in years.
balance: The current balance on the customer's credit card account.
products_credit_car: The number of credit card products the customer holds.
active_me: Whether the customer's account is currently active (1 for active, 0 for inactive).
estimated_churn: A binary variable indicating whether the customer is estimated to churn (1 for likely to churn, 0 otherwise).